Move a test to proper place
Its meaning has changed after another refactoring.
