+++ crates/air-lib/test-utils/Cargo.toml
@@ -26,5 +26,8 @@ once_cell = "1.15.0"
 semver = "1.0.14"
 serde_json = "1.0.86"
 
+[dev-dependencies]
+maplit = "1.0.2"
+
 [features]
 test_with_native_code = []
